requirements.txt
python-dotenv
google-generativeai
colorama